+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## v3.0.0 - 2021-09-06
+### Breaking changes
+- `wordsCount` does not longer hold a default value of `50`. You now need to explicitely provide a number greater than `0`.
+
+### Other changes
+- Add `charactersCount` option to give the flexibility to truncate the initial text content by characters count as well.
+
 ## v2.0.0 - 2021-08-20
 
 ### Breaking changes